requirements.txt
flask==2.3.3
flask-cors==4.0.0
openai-whisper==20231117
soundfile==0.12.1
numpy==1.24.3
torch==2.0.1 